requirements.txt
tensorflow
Pillow
numpy
tk
